pyproject.toml
[tool.poetry]
name = "lion-os"
version = "0.1.2"
description = ""
authors = ["HaiyangLi <quantocean.li@gmail.com>"]
readme = "README.md"
packages = [
    { include = "lion" },  # This ensures your package is still imported as 'lion'
]

[tool.poetry.dependencies]
python = "^3.11"
litellm = "^1.52.3"
pandas = "^2.2.3"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
httpx = "^0.27.2"
python-multipart = "^0.0.17"
psutil = "^6.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
ipykernel = "^6.29.5"
cryptography = "^43.0.3"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
